Move CardSampler to a test-only internal library
CardSampler is only used by the unit tests and benchmarks, but it was
shipped as part of the public API: its header lived in the installed
include/phevaluator tree and it was compiled into the pheval and
phevalplo{4,5,6} libraries.

Move it out of the published surface into a dedicated cpp/testutil
directory and build it as a standalone, non-installed static library
that both the tests and benchmarks link against. The header is no longer
installed, and it is dropped from the libraries' PUBLIC_HEADER lists and
source lists.

Copilot-Session: 07742674-f983-4ae2-a4d1-36972dad3598
