+++ cpp/CMakeLists.txt
@@ -19,7 +19,6 @@ option(BUILD_TESTS "Build test ON/OFF" ON)
 option(BUILD_EXAMPLES "Build examples ON/OFF" ON)
 
 add_library(pheval STATIC
-  src/card_sampler.cc
   src/dptables.c
   src/evaluator5.cc
   src/evaluator5.c
@@ -42,7 +41,6 @@ target_include_directories(pheval PUBLIC
 target_compile_options(pheval PUBLIC -O3)
 set(PUB_HEADERS include/phevaluator/phevaluator.h
                 include/phevaluator/card.h
-                include/phevaluator/card_sampler.h
                 include/phevaluator/rank.h)
 set_target_properties(pheval PROPERTIES
     VERSION ${PROJECT_VERSION}
@@ -116,7 +114,6 @@ endif()
 
 if (BUILD_PLO4)
   add_library(phevalplo4 STATIC
-    src/card_sampler.cc
     src/dptables.c
     src/evaluator_plo4.c
     src/evaluator_plo4.cc
@@ -133,7 +130,6 @@ if (BUILD_PLO4)
   target_compile_options(phevalplo4 PUBLIC -O3)
   set(PUB_HEADERS include/phevaluator/phevaluator.h
                   include/phevaluator/card.h
-                  include/phevaluator/card_sampler.h
                   include/phevaluator/rank.h)
   set_target_properties(phevalplo4 PROPERTIES
       VERSION ${PROJECT_VERSION}
@@ -152,7 +148,6 @@ if (BUILD_PLO5)
   )
 
   add_library(phevalplo5 STATIC
-    src/card_sampler.cc
     src/dptables.c
     src/evaluator_plo5.c
     src/evaluator_plo5.cc
@@ -169,7 +164,6 @@ if (BUILD_PLO5)
   target_compile_options(phevalplo5 PUBLIC -O3)
   set(PUB_HEADERS include/phevaluator/phevaluator.h
                   include/phevaluator/card.h
-                  include/phevaluator/card_sampler.h
                   include/phevaluator/rank.h)
   set_target_properties(phevalplo5 PROPERTIES
       VERSION ${PROJECT_VERSION}
@@ -187,7 +181,6 @@ if (BUILD_PLO6)
   )
 
   add_library(phevalplo6 STATIC
-    src/card_sampler.cc
     src/dptables.c
     src/evaluator_plo6.c
     src/evaluator_plo6.cc
@@ -204,7 +197,6 @@ if (BUILD_PLO6)
   target_compile_options(phevalplo6 PUBLIC -O3)
   set(PUB_HEADERS include/phevaluator/phevaluator.h
                   include/phevaluator/card.h
-                  include/phevaluator/card_sampler.h
                   include/phevaluator/rank.h)
   set_target_properties(phevalplo6 PROPERTIES
       VERSION ${PROJECT_VERSION}
@@ -230,6 +222,14 @@ if (BUILD_TESTS)
   )
   set(BENCHMARK_ENABLE_INSTALL OFF CACHE BOOL "" FORCE)
   FetchContent_MakeAvailable(googletest googlebenchmark)
+
+  # Internal helper used only by the tests and benchmarks below. It is
+  # intentionally not installed and its header is kept out of the public
+  # include/ tree so it is not part of the published library API.
+  add_library(card_sampler STATIC testutil/card_sampler.cc)
+  target_include_directories(card_sampler PUBLIC
+      ${CMAKE_CURRENT_SOURCE_DIR}/testutil)
+  target_compile_options(card_sampler PUBLIC -O3)
 
   if (BUILD_PLO4)
     set(unit_tests_source_plo4 test/evaluation_plo4.cc)
@@ -259,7 +259,7 @@ if (BUILD_TESTS)
     test/kev/kev_eval.c
   )
   target_compile_options(unit_tests PUBLIC -O3)
-  target_link_libraries(unit_tests gtest_main pheval
+  target_link_libraries(unit_tests gtest_main pheval card_sampler
     ${library_plo4}
     ${library_plo5}
     ${library_plo6}
@@ -274,7 +274,7 @@ if (BUILD_TESTS)
     ${benchmark_source_plo6}
   )
   target_compile_options(benchmark_phevaluator PUBLIC -O3)
-  target_link_libraries(benchmark_phevaluator benchmark pheval
+  target_link_libraries(benchmark_phevaluator benchmark pheval card_sampler
     ${library_plo4}
     ${library_plo5}
     ${library_plo6}
