+++ cpp/benchmark/benchmark.cc
@@ -1,6 +1,6 @@
 #include "benchmark/benchmark.h"
 
-#include "phevaluator/card_sampler.h"
+#include "card_sampler.h"
 #include "phevaluator/phevaluator.h"
 
 using namespace phevaluator;
